fixed checkstyle
